backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pypdfium2>=4.30.0
pymupdf>=1.24.0
azure-ai-documentintelligence>=1.0.2,<2.0
azure-ai-textanalytics>=5.3.0,<6.0
azure-cosmos>=4.6.0,<5.0
azure-storage-blob>=12.20.0,<13.0
azure-identity>=1.17.0,<2.0
openai>=1.102.0,<2.0
dependency-injector>=4.41.0
rapidfuzz>=3.0.0,<4.0
python-dotenv>=1.1.0
sse-starlette>=2.1.0,<3.0
pillow>=10.0.0,<12.0
aiohttp>=3.8.0,<4.0
agent-framework=1.0.0